… logstash-core/src/test/java/org/logstash/AccessorsTest.java