* `Accessors` with cached state get us nothing
   * Assuming a hot cache for a given path
       * For accessing top level fields they slow things down by adding another level of indirection
       * For accessing second level data they have the same level of indirection for `get`
       * sets destroy the cache for the whole `Event` => depending on the plugins, the cache can be almost useless
   * Assuming a cold cache the cache is only overhead obviously
   * Cache logic blows up the size of `Accessor` bytecode wise and prevents inlining into `Event`
   * The cache costs a lot of memory (with metrics turned off, this branch has an allocation byte rate about 25% below `master`) 
 * In practice with metrics turned off (metrics are too heavy, they gate everything when turned on at the moment by being a 50%+ performance hit on a maxed out box), this version of `Accessors` ~15-20% faster on 4 threads and the Apache dataset for `Xmx=1G` than `master` for me


Changes made:

* Removed redundant data and meta_data map references in `Accessors`
* Removed `lut` field that held a cache for `field reference => field value` used to avoid nested `Map<>` lookups
* This leaves `Accessors` with no state of its own => made its code path `static`
* This removes the ~50/50 (according to JitWatch) branching between paths hot and cold cache => the previous "cold" path is inlined better => split up all its sub-branches into individual methods to make the most of it and it seems now `Accessors` does in fact get fully inlined and eventually unloaded
* Also removed 100% redundant throws for IndexOutOfBoundsException and ClassCastException (if you look at both instances, those exceptions will be thrown automatically in all cases anyways, but `Throwable` constructors cannot be inlined 

Both of the below are 4 workers, 1024 batch size and metrics off at `Xmx = 1GB`.

### Before

<img width="2237" alt="screen shot 2017-07-28 at 23 52 16" src="https://user-images.githubusercontent.com/6490959/28737755-59b17f3c-73f0-11e7-8920-daa9f1fb234a.png">

### After

<img width="2230" alt="screen shot 2017-07-28 at 23 48 53" src="https://user-images.githubusercontent.com/6490959/28737759-5dcf8cb2-73f0-11e7-926f-101d9e5fa8c9.png">

* More minor, fewer major GCs
* Much lower allocation rate (see slope for old gen)

This should bring some help to issues with major GC like e.g. https://github.com/elastic/logstash/issues/5405#issuecomment-224840073

(if you let things run for a longer time than in the pictures above major pauses tend to be a lot shorter here too (I can run a few more examples tomorrow and add them here), which is expected since all event fields have much fewer paths to them now)

EDIT:

some benchmarks: 

Both of the below are 4 workers, 512 batch size and metrics off at `Xmx = 1GB`.

```sh
➜  logstash git:(lut-no-more) ✗ time cat ~/Downloads/apache_access_logs
13.2MiB 0:06:36 [34.0KiB/s] [                                                                                        <=>                                                                                        ]
 13801093
cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs  0.14s user 2.27s system 0% cpu 6:35.88 total
bin/logstash -b 512 -w 4 -f ~/tmp/logstash.cfg  1718.84s user 27.00s system 440% cpu 6:36.23 total
pv  2.66s user 16.51s system 4% cpu 6:36.23 total
wc -c  0.34s user 2.65s system 0% cpu 6:36.23 total
```

vs.

```sh
➜  logstash git:(master) ✗ time cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs | bin/logstash -b 512 -w 4 -f ~/tmp/logstash.cfg  | pv  | wc -c
13.2MiB 0:06:56 [32.4KiB/s] [       <=>                                                                                                                                                                        ]
 13801093
cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs  0.13s user 2.23s system 0% cpu 6:55.89 total
bin/logstash -b 512 -w 4 -f ~/tmp/logstash.cfg  1858.08s user 29.09s system 453% cpu 6:56.49 total
pv  3.42s user 21.46s system 5% cpu 6:56.49 total
wc -c  0.52s user 4.76s system 1% cpu 6:56.49 total
```

=> Less CPU use % and faster => GC wastes more cycles than caching saves.

But it's not only GC, turning `Xmx` up to `4G`:

```sh
➜  logstash git:(lut-no-more) ✗ time cat ~/Downloads/apache_access_logs
13.2MiB 0:05:59 [37.5KiB/s] [                                                                                                                                         <=>                                       ]
 13801157
cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs  0.13s user 2.22s system 0% cpu 5:59.22 total
bin/logstash -b 512 -w 4 -f ~/tmp/logstash.cfg  1524.02s user 25.50s system 430% cpu 5:59.84 total
```

```sh
➜  logstash git:(master) ✗ time cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs | bin/logstash -b 512 -w 4 -f ~/tmp/logstash.cfg  | pv  | wc -c
13.2MiB 0:06:26 [34.9KiB/s] [                                                                                                                    <=>                                                            ]
 13801157
cat ~/Downloads/apache_access_logs ~/Downloads/apache_access_logs  0.14s user 2.23s system 0% cpu 6:25.37 total
bin/logstash -b 512 -w 4 -f ~/tmp/logstash.cfg  1633.34s user 24.36s system 429% cpu 6:26.09 total
pv  2.69s user 17.03s system 5% cpu 6:26.09 total
wc -c  0.36s user 2.83s system 0% cpu 6:26.09 total
```

still faster, even when GC becomes trivial. Imo it's almost impossible to make a realistic case where this cache helps anything, given that is evicted on every `set` operation.
So the only winners here would be inefficient plugins that make redundant `get` calls to fields at nesting level 3 and up. Code without deep, redundant `get` calls can only loose imo.