@original-brownbear 2017-07-28
@suyograo sure will do tomorrow. Though it's not as interesting as runs with `pv | wc -c` because the tool needs the metrics and they way gate this at the moment :(
That said ... not to go too far off topic ... maybe the performance tool should also count the dots from the process output for those benchmark runs that have `dots` to `stdout` for better accuracy when comparing older versions (even if we fix the metrics, the metrics off baseline we can't compare with the tool otherwise).