@original-brownbear 2017-07-30
this gives rise to https://github.com/elastic/logstash/pull/7847 which comes with a significant reduction in `String` allocations and memory use