@guyboertje 2017-07-31
IIRC, the accessors implementation was a direct port of the Ruby Event + Accessors - there was little investigation into its efficacy - merely the path of least resistance to a Java Event that could serialize into the PQ.

Investigating this now, is due IMO.

**However, I think that other LS team members should understand what is changing and why. At the very least so that we can all be up to speed on such changes when handling Discuss, GH and SDH issues.**