@original-brownbear 2017-07-31
@guyboertje sure, I guess we can sent out an email once this is actually happening? Anything we should add that isn't in the PR description?